requirements.txt
pandas
statsmodels
lifelines
seaborn
matplotlib
scikit-learn
PyMuPDF
openpyxl
biopython
openai
requests
pypandoc
